# schemas/logweir-api-v1.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Logweir product API",
    "version": "1.0.0-alpha.1",
    "description": "The bounded shared-console product API (PLAT-17.1). Not a Kubernetes proxy: no generic resource path, no Secret, Pod, log, exec or Job access, no delete, and no arbitrary patch. Every error is application/problem+json. Unsafe methods require Content-Type: application/json and an Origin equal to the configured public origin. Durable creates require Idempotency-Key. Capabilities without a route are advertised false in /api/v1/session."
  },
  "paths": {
...
    "/api/v1/namespaces/{ns}/restores": {
      "get": {
        "operationId": "listRestores",
        "summary": "Restore projections with a status summary; planBytes omitted.",
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "description": "An explicitly granted namespace. An ungranted namespace is namespace_forbidden whether or not it exists.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size, 1 to 200. Default 50.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "The opaque nextCursor of the previous page. Bound to actor, route, namespace and filters; expires after 15 minutes.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labelSelector",
            "in": "query",
            "required": false,
            "description": "Equality-only label selector, key=value terms separated by commas.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreList"
                }
              }
            }
          },
          "default": {
            "description": "An application/problem+json error; the possible codes are listed in x-logweir-problem-codes.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-logweir-problem-codes": [
          "cursor_expired",
          "cursor_invalid",
          "forbidden",
          "header_not_allowed",
          "internal_error",
          "kubernetes_unavailable",
          "malformed_request",
          "method_not_allowed",
          "misdirected_request",
          "namespace_forbidden",
          "rate_limited",
          "session_expired",
          "unauthenticated",
          "upstream_timeout",
          "validation_failed"
        ]
      },
      "post": {
        "operationId": "createRestore",
        "summary": "Create a Restore. planBytes is opaque, checked against planHash and stored unchanged.",
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "description": "An explicitly granted namespace. An ungranted namespace is namespace_forbidden whether or not it exists.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "8 to 128 visible ASCII characters. The object name is derived from (actor, namespace, route, key); a replay of the same request returns 200 with the same object, a different request with the same key is idempotency_conflict.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "description": "Must equal the configured public origin exactly.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRestoreRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreResponse"
                }
              }
            }
          },
          "200": {
            "description": "Replayed an identical earlier request; the same object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreResponse"
                }
              }
            }
          },
          "default": {
            "description": "An application/problem+json error; the possible codes are listed in x-logweir-problem-codes.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-logweir-problem-codes": [
          "forbidden",
          "header_not_allowed",
          "idempotency_conflict",
          "idempotency_key_invalid",
          "idempotency_key_required",
          "internal_error",
          "kubernetes_unavailable",
          "malformed_request",
          "method_not_allowed",
          "misdirected_request",
          "namespace_forbidden",
          "origin_mismatch",
          "payload_too_large",
          "rate_limited",
          "session_expired",
          "state_conflict",
          "unauthenticated",
          "unsupported_media_type",
          "upstream_timeout",
          "validation_failed"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArchiveRequest": {
        "description": "An archive location in a request.",
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "credentialRef": {
            "description": "An existing Secret in this namespace carrying the object-store credential.",
            "default": null,
            "$ref": "#/components/schemas/NameRef",
            "nullable": true
          },
          "url": {
            "description": "`s3://`, `gs://`, `az://` or `file:///`; no userinfo.",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ArchiveView": {
        "description": "An archive location and the name of the Secret that reaches it.",
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "credentialRef": {
            "description": "The name of the Secret carrying the object-store credential. A name only; this service never reads Secrets.",
            "$ref": "#/components/schemas/NameRef",
            "nullable": true
          },
          "url": {
            "description": "The object-store URL, with any userinfo redacted.",
            "type": "string"
          }
        }
      },
...
      "CreateRestoreRequest": {
        "description": "`POST /api/v1/namespaces/{ns}/restores`.\n\n`planBytes` is OPAQUE: the service checks `planHash` against the SHA-256 of exactly these bytes and stores them unchanged. It never parses or re-emits the plan document.",
        "type": "object",
        "required": [
          "approvalRef",
          "backupSetRef",
          "deadlineSeconds",
          "planBytes",
          "planHash",
          "pointInTime",
          "sourceArchive",
          "target"
        ],
        "properties": {
          "approvalRef": {
            "description": "The Approval that will authorise this restore, in this namespace. It need not exist yet: the controller holds the restore until it verifies.",
            "$ref": "#/components/schemas/NameRef"
          },
          "backupSetRef": {
            "description": "The backup set ID inside the archive.",
            "type": "string"
          },
          "deadlineSeconds": {
            "description": "The Job deadline, 60 to 86400 seconds.",
            "type": "integer",
            "format": "int64"
          },
          "planBytes": {
            "description": "The exact plan document bytes the approval will bind, at most 256 KiB.",
            "type": "string"
          },
          "planHash": {
            "description": "`sha256:<lowercase hex>` of `planBytes`, as the client computed it.",
            "type": "string"
          },
          "pointInTime": {
            "description": "The RFC 3339 point in time to restore to.",
            "type": "string"
          },
          "sourceArchive": {
            "description": "The archive to restore from.",
            "$ref": "#/components/schemas/ArchiveRequest"
          },
          "target": {
            "description": "Where the restore writes.",
            "$ref": "#/components/schemas/RestoreTargetRequest"
          }
        },
        "additionalProperties": false
      },
...
      "FieldError": {
        "description": "One field-level validation failure.",
        "type": "object",
        "required": [
          "code",
          "field",
          "message"
        ],
        "properties": {
          "code": {
            "description": "A stable machine-readable code for this field, e.g. `required`.",
            "type": "string"
          },
          "field": {
            "description": "The JSON path of the field, e.g. `topics[2]`, or a header or query parameter name.",
            "type": "string"
          },
          "message": {
            "description": "A human-readable sentence. Never echoes a credential.",
            "type": "string"
          }
        }
      },
...
      "NameRef": {
        "description": "A same-namespace reference by name.",
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "description": "The referenced object's name, in the same namespace.",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "OperationSummary": {
        "description": "The status summary carried by list items.",
        "type": "object",
        "required": [
          "state",
          "terminal",
          "verificationState",
          "verifiedSuccess"
        ],
        "properties": {
          "state": {
            "description": "The lifecycle state.",
            "$ref": "#/components/schemas/OperationState"
          },
          "stateReason": {
            "description": "The CamelCase reason.",
            "type": "string",
            "nullable": true
          },
          "terminal": {
            "description": "Whether the state is final.",
            "type": "boolean"
          },
          "verificationState": {
            "description": "The verification state.",
            "$ref": "#/components/schemas/VerificationState"
          },
          "verifiedSuccess": {
            "description": "Result pass AND verification `Valid`.",
            "type": "boolean"
          }
        }
      },
...
      "Page": {
        "description": "One page of a list.",
        "type": "object",
        "required": [
          "limit"
        ],
        "properties": {
          "limit": {
            "description": "The page size that was applied.",
            "type": "integer",
            "format": "uint32",
            "minimum": 0.0
          },
          "nextCursor": {
            "description": "An opaque cursor for the next page, valid for fifteen minutes for the same actor, route, namespace and filters; absent on the last page.",
            "type": "string",
            "nullable": true
          },
          "snapshot": {
            "description": "The Kubernetes list resourceVersion this page was read at.",
            "type": "string",
            "nullable": true
          }
        }
      },
      "Problem": {
        "description": "The rendered problem document.",
        "type": "object",
        "required": [
          "code",
          "detail",
          "requestId",
          "retryable",
          "status",
          "title",
          "type"
        ],
        "properties": {
          "code": {
            "description": "The stable code.",
            "$ref": "#/components/schemas/ProblemCode"
          },
          "detail": {
            "description": "A sentence this service wrote. Never a Kubernetes message.",
            "type": "string"
          },
          "errors": {
            "description": "Field-level failures, when the code is `validation_failed`.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "requestId": {
            "description": "The request ID, also returned as `X-Request-ID`.",
            "type": "string"
          },
          "retryable": {
            "description": "Whether repeating the identical request later can succeed.",
            "type": "boolean"
          },
          "status": {
            "description": "The HTTP status, repeated.",
            "type": "integer",
            "format": "uint16",
            "minimum": 0.0
          },
          "title": {
            "description": "A fixed title for the code.",
            "type": "string"
          },
          "type": {
            "description": "`https://logweir.dev/problems/<code>`.",
            "type": "string"
          }
        }
      },
      "ProblemCode": {
        "description": "The stable problem codes. Clients branch on `code`, never on `detail`.\n\nThe decision's baseline set is here verbatim (`unauthenticated`, `session_expired`, `forbidden`, `namespace_forbidden`, `not_found`, `idempotency_conflict`, `state_conflict`, `approval_required`, `policy_mismatch`, `precondition_failed`, `validation_failed`, `rate_limited`, `kubernetes_unavailable`, `upstream_timeout`, plus the cursor pair). The remainder name transport-level refusals the baseline did not enumerate; each is documented in the OpenAPI document.",
        "oneOf": [
          {
            "description": "400 — the body is not JSON, a query string is malformed or carries an unknown or repeated parameter.",
            "type": "string",
            "enum": [
              "malformed_request"
            ]
          },
          {
            "description": "400 — a header this service refuses outright (`Impersonate-*`).",
            "type": "string",
            "enum": [
              "header_not_allowed"
            ]
          },
          {
            "description": "400 — a durable POST arrived without `Idempotency-Key`.",
            "type": "string",
            "enum": [
              "idempotency_key_required"
            ]
          },
          {
            "description": "400 — `Idempotency-Key` is not 8–128 visible ASCII characters, or is present on a route that does not accept it.",
            "type": "string",
            "enum": [
              "idempotency_key_invalid"
            ]
          },
          {
            "description": "400 — a list cursor failed authentication or names another scope.",
            "type": "string",
            "enum": [
              "cursor_invalid"
            ]
          },
          {
            "description": "401 — no authenticated actor (PLAT-17.2 authenticators).",
            "type": "string",
            "enum": [
              "unauthenticated"
            ]
          },
          {
            "description": "401 — the session is past its signed expiry (PLAT-17.2).",
            "type": "string",
            "enum": [
              "session_expired"
            ]
          },
          {
            "description": "403 — the actor is not permitted this action.",
            "type": "string",
            "enum": [
              "forbidden"
            ]
          },
          {
            "description": "403 — the namespace is not explicitly granted to the actor.",
            "type": "string",
            "enum": [
              "namespace_forbidden"
            ]
          },
          {
            "description": "403 — an unsafe method without an `Origin` exactly equal to the configured public origin.",
            "type": "string",
            "enum": [
              "origin_mismatch"
            ]
          },
          {
            "description": "404 — no such route or no such object.",
            "type": "string",
            "enum": [
              "not_found"
            ]
          },
          {
            "description": "405 — the route exists and does not accept this method.",
            "type": "string",
            "enum": [
              "method_not_allowed"
            ]
          },
          {
            "description": "409 — the idempotency key was already used with a different request.",
            "type": "string",
            "enum": [
              "idempotency_conflict"
            ]
          },
          {
            "description": "409 — the deterministic object exists but was not created by this idempotency scope; it is never adopted.",
            "type": "string",
            "enum": [
              "state_conflict"
            ]
          },
          {
            "description": "409 — an approval is required first (PLAT-19.2).",
            "type": "string",
            "enum": [
              "approval_required"
            ]
          },
          {
            "description": "409 — the bound approval policy does not match (PLAT-19.2).",
            "type": "string",
            "enum": [
              "policy_mismatch"
            ]
          },
          {
            "description": "410 — the list cursor expired or Kubernetes compacted its continue token; restart the list without a cursor.",
            "type": "string",
            "enum": [
              "cursor_expired"
            ]
          },
          {
            "description": "412 — `expectedResourceVersion` is not the object's current version.",
            "type": "string",
            "enum": [
              "precondition_failed"
            ]
          },
          {
            "description": "413 — the request body exceeds its limit.",
            "type": "string",
            "enum": [
              "payload_too_large"
            ]
          },
          {
            "description": "415 — an unsafe method without `Content-Type: application/json`.",
            "type": "string",
            "enum": [
              "unsupported_media_type"
            ]
          },
          {
            "description": "421 — the `Host` header names an authority this listener does not serve (DNS-rebinding defence).",
            "type": "string",
            "enum": [
              "misdirected_request"
            ]
          },
          {
            "description": "422 — the request is well-formed JSON and fails validation.",
            "type": "string",
            "enum": [
              "validation_failed"
            ]
          },
          {
            "description": "429 — Kubernetes throttled the request; see `Retry-After`.",
            "type": "string",
            "enum": [
              "rate_limited"
            ]
          },
          {
            "description": "500 — an invariant this service holds did not hold.",
            "type": "string",
            "enum": [
              "internal_error"
            ]
          },
          {
            "description": "503 — Kubernetes could not be reached or refused this service's own identity.",
            "type": "string",
            "enum": [
              "kubernetes_unavailable"
            ]
          },
          {
            "description": "504 — the Kubernetes call exceeded its 10-second deadline.",
            "type": "string",
            "enum": [
              "upstream_timeout"
            ]
          }
        ]
      },
...
      "Restore": {
        "description": "A restore run.",
        "type": "object",
        "required": [
          "approvalRef",
          "backupSetRef",
          "deadlineSeconds",
          "name",
          "namespace",
          "newTopics",
          "operation",
          "planBytesLength",
          "planHash",
          "pointInTime",
          "resourceVersion",
          "sourceArchive",
          "target",
          "uid"
        ],
        "properties": {
          "approvalRef": {
            "description": "The Approval name this restore waits for.",
            "$ref": "#/components/schemas/NameRef"
          },
          "backupSetRef": {
            "description": "The backup set ID.",
            "type": "string"
          },
          "createdAt": {
            "description": "When the object was created.",
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deadlineSeconds": {
            "description": "The Job deadline.",
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "description": "The object name.",
            "type": "string"
          },
          "namespace": {
            "description": "The namespace.",
            "type": "string"
          },
          "newTopics": {
            "description": "The topics this run created.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "operation": {
            "description": "The normalized status summary.",
            "$ref": "#/components/schemas/OperationSummary"
          },
          "planBytes": {
            "description": "The stored plan bytes, unchanged. Present on the single-object read only; lists omit it.",
            "type": "string",
            "nullable": true
          },
          "planBytesLength": {
            "description": "The stored plan bytes' length in bytes.",
            "type": "integer",
            "format": "uint",
            "minimum": 0.0
          },
          "planHash": {
            "description": "`sha256:<hex>` of the stored plan bytes, computed by this service.",
            "type": "string"
          },
          "pointInTime": {
            "description": "The point in time.",
            "type": "string",
            "format": "date-time"
          },
          "resourceVersion": {
            "description": "The resourceVersion.",
            "type": "string"
          },
          "sourceArchive": {
            "description": "The archive restored from.",
            "$ref": "#/components/schemas/ArchiveView"
          },
          "target": {
            "description": "Where it writes.",
            "$ref": "#/components/schemas/RestoreTargetView"
          },
          "uid": {
            "description": "The Kubernetes UID.",
            "type": "string"
          }
        }
      },
      "RestoreList": {
        "type": "object",
        "required": [
          "items",
          "page",
          "requestId"
        ],
        "properties": {
          "items": {
            "description": "The items on this page.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Restore"
            }
          },
          "page": {
            "description": "Paging.",
            "$ref": "#/components/schemas/Page"
          },
          "requestId": {
            "description": "The request ID.",
            "type": "string"
          }
        }
      },
...
      "RestoreResponse": {
        "type": "object",
        "required": [
          "item",
          "requestId"
        ],
        "properties": {
          "item": {
            "description": "The item.",
            "$ref": "#/components/schemas/Restore"
          },
          "replayed": {
            "description": "On a durable create: whether this response replays an earlier identical request (HTTP 200) rather than creating (HTTP 201).",
            "type": "boolean",
            "nullable": true
          },
          "requestId": {
            "description": "The request ID.",
            "type": "string"
          }
        }
      },
      "RestoreTargetRequest": {
        "description": "Where a restore writes.",
        "type": "object",
        "required": [
          "clusterRef",
          "mode",
          "topicNaming"
        ],
        "properties": {
          "clusterRef": {
            "description": "The target connection, in this namespace.",
            "$ref": "#/components/schemas/NameRef"
          },
          "mode": {
            "description": "`scratch` or `newTopic`.",
            "$ref": "#/components/schemas/RestoreMode"
          },
          "topicNaming": {
            "description": "Topic naming.",
            "$ref": "#/components/schemas/TopicNamingRequest"
          }
        },
        "additionalProperties": false
      },
      "RestoreTargetView": {
        "description": "A restore target, as stored.",
        "type": "object",
        "required": [
          "clusterRef",
          "mode",
          "topicPrefix"
        ],
        "properties": {
          "clusterRef": {
            "description": "The target connection.",
            "$ref": "#/components/schemas/NameRef"
          },
          "mode": {
            "description": "`scratch` or `newTopic`.",
            "$ref": "#/components/schemas/RestoreMode"
          },
          "topicPrefix": {
            "description": "The topic prefix.",
            "type": "string"
          }
        }
      },
...
    }
  },
...
}